Skapat en component med navbar i för alla pages.
