+++ lingoquest-web/src/App.js
@@ -3,12 +3,13 @@ import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 import Home from './pages/home';
 import Checklist from './pages/checklist';
 import './App.css';
+import Navbar from './components/navbar';
 
 function App() {
   return (
     <Router>
       <div>
-        {/* Define the routes */}
+        <Navbar/>
         <Routes>
           <Route path="/" element={<Home />} />
           <Route path="/checklist" element={<Checklist />} />
